Class3/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pyttsx3==2.90
pydub==0.25.1
python-dotenv==1.0.1
transformers==4.42.4
accelerate==0.33.0
safetensors==0.4.3
sentencepiece==0.2.0
openai-whisper==20231117
# NOTE: Install torch separately per platform from https://pytorch.org/get-started/locally/
